+++ src/System/include/BipedalLocomotion/System/SharedResource.h
@@ -21,7 +21,9 @@ namespace System
  * shared resource you can use it only with shared_ptr. For this reason the only way to create a
  * shared resource is to use the SharedResource::create() static method. If want a visual
  * interpretation of shared resource you may imagine it as the wire connecting two blocks running in
- * two separate threads. The implementation of SharedResource is thread safe.
+ * two separate threads. The implementation of SharedResource is thread safe, except when getting the
+ * resource by reference or when the resource type is a smart pointer.
+ * In this case, the user is obliged to access the mutex and lock/unlock it as required.
  */
 template <class T> class SharedResource
 {
@@ -45,6 +47,20 @@ public:
     T get();
 
     /**
+     * Get the resource using const reference.
+     * @note one must not forget to access the mutex and
+     *       perform necessary locking and unlocking
+     * @return the const reference of the object inside the shared resource.
+     */
+    const T& get() const;
+
+    /**
+     * Get the mutex by reference.
+     * @return the reference of the mutex.
+     */
+    std::mutex& mutex();
+
+    /**
      * Method used to create a shared resource.
      * @return a pointer of a shared resource.
      */
@@ -61,6 +77,16 @@ template <class T> T SharedResource<T>::get()
 {
     const std::lock_guard<std::mutex> lock(m_mutex);
     return m_resource;
+}
+
+template <class T> const T& SharedResource<T>::get() const
+{
+    return m_resource;
+}
+
+template <class T> std::mutex& SharedResource<T>::mutex()
+{
+    return m_mutex;
 }
 
 template <class T> typename SharedResource<T>::Ptr SharedResource<T>::create()